FruUtils: Avoid truncation in derivation of fruAreaSize
FRU areas can be up to 255 * 8 bytes in length. 2040 can't be
represented in an unsigned 8-bit type, so widen the result type to
match fruBlockSize.

+++ src/FruUtils.cpp
@@ -304,7 +304,7 @@ resCodes formatFRU(const std::vector<uint8_t>& fruBytes,
             return resCodes::resErr;
         }
 
-        uint8_t fruAreaSize = *fruBytesIter * fruBlockSize;
+        size_t fruAreaSize = *fruBytesIter * fruBlockSize;
         std::vector<uint8_t>::const_iterator fruBytesIterEndArea =
             fruBytes.begin() + offset + fruAreaSize - 1;
         ++fruBytesIter;
